Added discount,selling and purchase in insert

+++ OP_Inventory_Billing_App/lib/widgets/billing_card.dart
@@ -5,7 +5,7 @@ import 'TextWidget.dart';
 class BillingCard extends StatelessWidget {
 
   final String name;
-  final int cost;
+  final double cost;
   final int discount;
   final String time;
   final int items;
